feat: widen integer literals from u64 to u128, enable 65-bit CHERI init
The Tourbillon compiler stored integer literals as u64, which truncated
the tag bit (bit 64) of 65-bit capability init values. This forced a
testbench hack to patch the PCC tag after reset.

Widen Literal::Int, Token::Int, init_tokens, and Cell init to u128
throughout the compiler pipeline (lexer → parser → AST → IR → lower).
The next_pcc FIFO now initializes with the full root PCC including tag=1:
  init = 0x1FFC0818080000000

Also adds bounds checking to Execute (Phase 3a): loads/stores through
tagged capabilities are checked against bounds and permissions.
Violations suppress the memory access and zero the destination (soft trap).
Tag=0 caps skip checks for backward compatibility with rv32ui tests.

+++ src/ast.rs
@@ -264,7 +264,7 @@ pub enum Expr {
 
 #[derive(Debug, Clone)]
 pub enum Literal {
-    Int(u64),
+    Int(u128),
     Bool(bool),
 }
 
@@ -327,7 +327,7 @@ pub struct QueueDecl {
     pub depth: Option<u64>,
     /// Number of initial tokens pre-loaded at reset.
     /// `init = N` in pipe declarations; enables deadlock-free cycles.
-    pub init_tokens: Option<u64>,
+    pub init_tokens: Option<u128>,
     /// `external Queue(...)` — no FIFO instantiated, signals become module ports.
     pub is_external: bool,
 }
